Made accountMaster from Long to int

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/services/payment/Payment.java
@@ -51,7 +51,7 @@ public interface Payment {
 	 * @return
 	 * @throws NonFatalException
 	 */
-	public boolean subscribe(User user,Company company, Long planId, String nonce) throws InvalidInputException, PaymentException;
+	public boolean subscribe(User user,Company company, int planId, String nonce) throws InvalidInputException, PaymentException;
 	
 	/**
 	 * Function to create a Braintree transaction with a particular payment method token and an amount
